Added searchRoute, Fuse.js autocomplete and updated player/team/stadium services

+++ Controller/teamController.js
@@ -6,6 +6,17 @@ const News = require("../models/newsModel");
 const Player = require("./../models/playerModel");
 const { options } = require("../app");
 
+exports.createTeam = catchAsync(async (req, res, next) => {
+  const team = await Team.findOne({ name: req.body.name });
+  if (team) return next(new appError("team already exist", 400));
+  const newTeam = await Team.create(req.body);
+  res.status(201).json({
+    status: "success",
+    message: "team created Successfully",
+    data: { newTeam },
+  });
+});
+
 //Get All Teams
 exports.getAllTeams = catchAsync(async (req, res, next) => {
   const teams = await Team.find();
@@ -21,6 +32,8 @@ exports.getAllTeams = catchAsync(async (req, res, next) => {
 exports.getOneTeam = catchAsync(async (req, res, next) => {
   const teamId = req.params.id;
   if (!teamId) return next(new appError("please enter teamId", 400));
+  if (!mongoose.Types.ObjectId.isValid(teamId))
+    return next(new appError("id not valid", 400));
   const team = await Team.findById(teamId);
   if (!team) return next(new appError("not exist", 404));
   res.status(200).json({
@@ -29,18 +42,11 @@ exports.getOneTeam = catchAsync(async (req, res, next) => {
   });
 });
 
-//Create Team By Admin
-exports.createTeam = catchAsync(async (req, res, next) => {
-  const team = await Team.create(req.body);
-  res.status(201).json({
-    status: "success",
-    data: { team },
-  });
-});
-
 //Update Team By Admin
 exports.updateTeam = catchAsync(async (req, res, next) => {
   const teamId = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(teamId))
+    return next(new appError("id not valid", 400));
   const team = await Team.findByIdAndUpdate(teamId, req.body, {
     new: true,
     runValidators: true,
@@ -48,6 +54,7 @@ exports.updateTeam = catchAsync(async (req, res, next) => {
   if (!team) return next(new appError("not exist", 404));
   res.status(200).json({
     status: "success",
+    message: "team updated successfully",
     data: { team },
   });
 });
@@ -55,6 +62,8 @@ exports.updateTeam = catchAsync(async (req, res, next) => {
 //Delete Team By Admin
 exports.deleteTeam = catchAsync(async (req, res, next) => {
   const teamId = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(teamId))
+    return next(new appError("id not valid", 400));
   const team = await Team.findByIdAndDelete(teamId);
   if (!team) return next(new appError("not exist", 404));
   res.status(204).json({
@@ -80,56 +89,5 @@ exports.informationOfTeam = catchAsync(async (req, res, next) => {
   res.status(200).json({
     status: "success",
     data: { information: team.information },
-  });
-});
-
-exports.addCoachToTeam = catchAsync(async (req, res, next) => {
-  const teamId = req.params.id;
-  if (!teamId) return next(new appError("please enter teamId", 400));
-  const { coachId } = req.body;
-  if (!coachId) return next(new appError("please enter coachId", 400));
-  const team = await Team.findByIdAndUpdate(
-    teamId,
-    {
-      $addToSet: { coach: coachId },
-    },
-    { new: true },
-  ).populate({ path: "coach", select: "name avatar" });
-  if (!team) return next(new appError("not exist", 404));
-  res.status(200).json({
-    status: "success",
-    data: { team },
-  });
-});
-exports.getCoachInTeam = catchAsync(async (req, res, next) => {
-  const teamId = req.params.id;
-  if (!teamId) return next(new appError("please enter teamId", 400));
-  const team = await Team.findById(teamId).populate({
-    path: "coach",
-    select: "name avatar",
-  });
-  if (!team) return next(new appError("not exist", 404));
-  res.status(200).json({
-    status: "success",
-    data: { coach: team.coach },
-  });
-});
-exports.deleteCoachFromTeam = catchAsync(async (req, res, next) => {
-  const teamId = req.params.id;
-  if (!teamId) return next(new appError("please enter teamId", 400));
-  const { coachId } = req.body;
-  if (!coachId) return next(new appError("please enter coachId", 400));
-  const team = await Team.findByIdAndUpdate(
-    teamId,
-    {
-      $pull: { coach: coachId },
-    },
-    { new: true },
-  ).populate({ path: "coach", select: "name avatar" });
-  if (!team) return next(new appError("not exist", 404));
-  res.status(200).json({
-    status: "success",
-    message: "coach deleted successfully",
-    data: { coach: team.coach },
   });
 });
